Add features
Add new power for towers in last upgrade:
AOE area for bomb tower
Periodic damage for fire tower

Create health bar for enemy
Cleanup code and refactor enemy system

+++ Assets/Scripts/EventController.cs
@@ -6,6 +6,7 @@ public class EventController : MonoBehaviour
 {
     public static UnityEvent OnEnemyDestroy { get; set; } = new UnityEvent();
     public static UnityEvent<int> OnEnemyCoinsAmount { get; set; } = new UnityEvent<int>();
+
     public static UnityEvent<int> OnCastleTakeDamage { get; set; } = new UnityEvent<int>();
 
     public static UnityEvent OnGameOver { get; set; } = new UnityEvent();
@@ -15,7 +16,7 @@ public class EventController : MonoBehaviour
     public static UnityEvent<int> OnTowerSell { get; set; } = new UnityEvent<int>();
     public static UnityEvent<int> OnTowerUpgrade { get; set; } = new UnityEvent<int>();
 
-    public static UnityEvent<int, int, int> OnUpdateInfoUI { get; set; } = new UnityEvent<int, int, int>();
+    public static UnityEvent<float, float, float> OnUpdateInfoUI { get; set; } = new UnityEvent<float, float, float>();
     public static UnityEvent<TowerController> OnUpdateButtonsUI { get; set; } = new UnityEvent<TowerController>();
 
     public static UnityEvent<PlotTowerSettings> OnPlotSelected { get; set; } = new UnityEvent<PlotTowerSettings>();
